Register names for the autogenerated dhcp addresses
This is for the `dhcpd` address method used for Ethernet connections
that only have two computers on the LAN. The most common use of this is
when connecting a device to a laptop via a virtual Ethernet interface on
a USB cable. A DHCP server is run on the device that assigns the laptop
an address. Since the assigned address is computed, programs can't
hardcode one in to refer to the laptop. This change adds `peer.usb0.lan`
and `local.usb0.lan` to the Erlang resolver's host table where `usb0` is
the Ethernet interface that's being used.

You can query the addresses using `:inet.gethostbyname/1` or dump the
host table using `:inet.get_rc/0`.